+++ queued_storage/test_settings.py
@@ -11,7 +11,6 @@ INSTALLED_APPS = [
     'djcelery',
     'queued_storage',
     'queued_storage.tests',
-    'django_jenkins',
 ]
 
 import djcelery
@@ -22,9 +21,4 @@ CELERY_IGNORE_RESULT = True
 CELERYD_LOG_LEVEL = "DEBUG"
 CELERY_DEFAULT_QUEUE = "queued_storage"
 
-JENKINS_TASKS = (
-    'django_jenkins.tasks.run_pyflakes',
-    'django_jenkins.tasks.run_pep8',
-    'django_jenkins.tasks.with_coverage',
-    'django_jenkins.tasks.django_tests',
-)
+TEST_RUNNER = 'discover_runner.DiscoverRunner'