+++ docs/changelog.rst
@@ -1,6 +1,26 @@
 Changelog
 =========
 
+v0.6 (2012-05-24)
+-----------------
+
+- Added `file_transferred` signal that is called right after a file has been
+  transfered from the local to the remote storage.
+
+- Switched to using `django-discover-runner`_ and Travis for testing:
+  http://travis-ci.org/jezdez/django-queued-storage
+
+.. `django-discover-runner`: http://pypi.python.org/pypi/django-discover-runner
+
+v0.5 (2012-03-19)
+-----------------
+
+- Fixed retrying in case of errors.
+
+- Dropped Python 2.5 support as Celery has dropped it, too.
+
+- Use django-jenkins.
+
 v0.4 (2011-11-03)
 -----------------
 
@@ -20,7 +40,7 @@ v0.4 (2011-11-03)
 
       my_storage = QueuedStorage(
           FileSystemStorage(location='/path/to/files'),
-      	  MyCustomStorageBackend(spam='eggs'))
+          MyCustomStorageBackend(spam='eggs'))
 
   New::
 
@@ -32,8 +52,8 @@ v0.4 (2011-11-03)
 
   .. warning::
 
-  	This change is backwards-incompatible if you used the
-  	:class:`~queued_storage.backends.QueuedStorage` API.
+     This change is backwards-incompatible if you used the
+     :class:`~queued_storage.backends.QueuedStorage` API.
 
 v0.3 (2011-09-19)
 -----------------